Add first UEFI support
- Remove defineVariables() function
The variables are set in the end of the script, in the global scope.

- Add UEFI flag
This flag tells the script to make an uefi-bootable media

- Update help text
The text of help shows and explains the new UEFI flag

- Update version number

- Update verbose messages
The messages are cleaner, I removed the dots to improve readability and
I put variables inside the messages to make it easier for us to debug.